requirements.txt
pillow
exifread